requirements.txt
flask
flask-caching
flask-cors
python-dotenv